refactor: 배포시 고친 에러 재수정
